Display List Abstractions, Commands Working
I have abstracted some more of the display list commands. I also
reviewed the programming model for the command list and implement
it according to the ring buffer in the datasheet for the FT800.

There is a functional demo in main.c that moves a square across the
screen from top to bottom, left to right.