Driver Updates, Text Rendering, Numbers Rendered
I have changed the driver to use the coprocessor for all commands.
This allows a mix of high level and low level display list commands.

I implemented the text rendering command and used this to render
numbers on the u2048 game board.

I have created a SysCalls.c file. Newlib was complaining about missing
_sbrk when I attempted to use sprintf. I have defined it to always
return 0. This will probably bite me later, but it fixed the undefined
reference. Time to learn how to compile newlib and get that whole thing
online.